feat: separate next config and next-18n config
