--- next.config.js
@@ -1,15 +1,0 @@
-/** @type {import('next').NextConfig} */
-
-import path from 'path';
-
-const nextConfig = {
-  reactStrictMode: true,
-
-  i18n: {
-    defaultLocale: 'en',
-    locales: ['en', 'ru'],
-    localePath: path.resolve('./public/locales'),
-  },
-};
-
-module.exports = nextConfig;
