+++ tsconfig.json
@@ -25,7 +25,8 @@
     "**/*.ts",
     "**/*.tsx",
     "lint-staged.config.js",
-    "next.config.js",
+    "next-i18next.config.mjs",
+    "next.config.mjs",
     "commitlint.config.js"
   ],
   "exclude": ["node_modules"]
